# Had Conflicts in those files:
#	addons/resource.language.en_gb/resources/strings.po
#	system/shaders/GL/1.2/gl_shader_vert_simple.glsl
#	system/shaders/GL/1.5/gl_shader_vert_simple.glsl
#	tools/depends/target/pythonmodule-pil/PYTHONMODULE-PIL-VERSION
#	xbmc/settings/MediaSettings.cpp
#	xbmc/settings/Settings.cpp
#	xbmc/video/dialogs/GUIDialogAudioSettings.cpp
#	xbmc/video/dialogs/GUIDialogAudioSettings.h
#	xbmc/windows/GUIWindowSystemInfo.cpp
#	xbmc/windows/GUIWindowSystemInfo.h

+++ xbmc/pvr/guilib/PVRGUIChannelIconUpdater.cpp
@@ -65,7 +65,7 @@ void CPVRGUIChannelIconUpdater::SearchAndUpdateMissingChannelIcons() const
   for (const auto& group : m_groups)
   {
     const std::vector<std::shared_ptr<CPVRChannelGroupMember>> members = group->GetMembers();
-    int channelIndex = 0;
+    size_t channelIndex = 0;
     for (const auto& member : members)
     {
       const std::shared_ptr<CPVRChannel> channel = member->Channel();
